requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
google-generativeai
google-cloud-bigquery
google-cloud-bigquery-storage
google-auth
typer[all]
rich
pydantic

langgraph
langchain-core
